# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BOARDS - Basketball Operations, Analysis & R&D System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: number;
  firstName: string;
  lastName: string;
  position: string;
  jerseyNumber: number;
  teamId: number;
  /** @nullable */
  height?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  graduationYear?: number | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PlayerInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  position: string;
  jerseyNumber: number;
  teamId: number;
  height?: string;
  weight?: number;
  graduationYear?: number;
  status?: string;
  notes?: string;
}

export interface PlayerUpdate {
  /** @minLength 1 */
  firstName?: string;
  /** @minLength 1 */
  lastName?: string;
  position?: string;
  jerseyNumber?: number;
  teamId?: number;
  height?: string;
  weight?: number;
  graduationYear?: number;
  status?: string;
  notes?: string;
}

export interface Team {
  id: number;
  name: string;
  level: string;
  season: string;
  record: string;
  /** @nullable */
  coachName?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface TeamInput {
  /** @minLength 1 */
  name: string;
  level: string;
  season: string;
  record?: string;
  coachName?: string;
  notes?: string;
}

export interface TeamUpdate {
  /** @minLength 1 */
  name?: string;
  level?: string;
  season?: string;
  record?: string;
  coachName?: string;
  notes?: string;
}

export interface Practice {
  id: number;
  teamId: number;
  scheduledAt: string;
  durationMinutes: number;
  focus: string;
  /** @nullable */
  drills?: string | null;
  /** @nullable */
  notes?: string | null;
  status: string;
  createdAt: string;
}

export interface PracticeInput {
  teamId: number;
  scheduledAt: string;
  durationMinutes: number;
  focus: string;
  drills?: string;
  notes?: string;
  status?: string;
}

export interface PracticeUpdate {
  scheduledAt?: string;
  durationMinutes?: number;
  focus?: string;
  drills?: string;
  notes?: string;
  status?: string;
}

export interface Game {
  id: number;
  teamId: number;
  opponent: string;
  scheduledAt: string;
  location: string;
  isHome?: boolean;
  result: string;
  /** @nullable */
  scoreUs?: number | null;
  /** @nullable */
  scoreThem?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface GameInput {
  teamId: number;
  /** @minLength 1 */
  opponent: string;
  scheduledAt: string;
  location: string;
  isHome?: boolean;
  result?: string;
  scoreUs?: number;
  scoreThem?: number;
  notes?: string;
}

export interface GameUpdate {
  /** @minLength 1 */
  opponent?: string;
  scheduledAt?: string;
  location?: string;
  isHome?: boolean;
  result?: string;
  scoreUs?: number;
  scoreThem?: number;
  notes?: string;
}

export interface ScoutingReport {
  id: number;
  teamId: number;
  opponentName: string;
  /** @nullable */
  gameDate?: string | null;
  /** @nullable */
  offensiveSystem?: string | null;
  /** @nullable */
  defensiveSystem?: string | null;
  /** @nullable */
  keyPlayers?: string | null;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  weaknesses?: string | null;
  /** @nullable */
  adjustments?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ScoutingReportInput {
  teamId: number;
  /** @minLength 1 */
  opponentName: string;
  gameDate?: string;
  offensiveSystem?: string;
  defensiveSystem?: string;
  keyPlayers?: string;
  strengths?: string;
  weaknesses?: string;
  adjustments?: string;
  notes?: string;
}

export interface ScoutingReportUpdate {
  /** @minLength 1 */
  opponentName?: string;
  gameDate?: string;
  offensiveSystem?: string;
  defensiveSystem?: string;
  keyPlayers?: string;
  strengths?: string;
  weaknesses?: string;
  adjustments?: string;
  notes?: string;
}

export interface DashboardSummary {
  totalPlayers: number;
  totalTeams: number;
  upcomingPractices: number;
  upcomingGames: number;
  totalWins: number;
  totalLosses: number;
  scoutingReports: number;
}

export interface UpcomingEvent {
  id: number;
  type: string;
  title: string;
  scheduledAt: string;
  teamName: string;
  /** @nullable */
  location?: string | null;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  createdAt: string;
}