# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BOARDS - Basketball Operations, Analysis & R&D System API
 * OpenAPI spec version: 0.1.0
 */

export interface Game {
  id: number;
  teamId: number;
  opponent: string;
  scheduledAt: string;
  location: string;
  isHome?: boolean;
  result: string;
  /** @nullable */
  scoreUs?: number | null;
  /** @nullable */
  scoreThem?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}