# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BOARDS - Basketball Operations, Analysis & R&D System API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  firstName: string;
  lastName: string;
  position: string;
  jerseyNumber: number;
  teamId: number;
  /** @nullable */
  height?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  graduationYear?: number | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}