# lib/api-zod/src/generated/types/practice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BOARDS - Basketball Operations, Analysis & R&D System API
 * OpenAPI spec version: 0.1.0
 */

export interface Practice {
  id: number;
  teamId: number;
  scheduledAt: string;
  durationMinutes: number;
  focus: string;
  /** @nullable */
  drills?: string | null;
  /** @nullable */
  notes?: string | null;
  status: string;
  createdAt: string;
}