# lib/api-zod/src/generated/types/practiceInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BOARDS - Basketball Operations, Analysis & R&D System API
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeInput {
  teamId: number;
  scheduledAt: string;
  durationMinutes: number;
  focus: string;
  drills?: string;
  notes?: string;
  status?: string;
}