# lib/api-zod/src/generated/types/practiceUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BOARDS - Basketball Operations, Analysis & R&D System API
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeUpdate {
  scheduledAt?: string;
  durationMinutes?: number;
  focus?: string;
  drills?: string;
  notes?: string;
  status?: string;
}