# lib/api-zod/src/generated/types/scoutingReportInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BOARDS - Basketball Operations, Analysis & R&D System API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoutingReportInput {
  teamId: number;
  /** @minLength 1 */
  opponentName: string;
  gameDate?: string;
  offensiveSystem?: string;
  defensiveSystem?: string;
  keyPlayers?: string;
  strengths?: string;
  weaknesses?: string;
  adjustments?: string;
  notes?: string;
}